requirements.txt
numpy==2.0.2
opencv-python==4.11.0.86
gradio==5.16.1
requests==2.32.3
mysql-connector-python==9.2.0
matplotlib==3.10.0